+++ test/model.geosearch.test.js
@@ -27,6 +27,8 @@ describe('model', function() {
   });
 
   describe('geoSearch', function() {
+    this.timeout(process.env.TRAVIS ? 8000 : 4500);
+
     it('works', function(done) {
       var Geo = getModel(db);
       assert.ok(Geo.geoSearch instanceof Function);